requirements.txt
typer==0.21.1
duckdb==1.4.3
pyyaml>=6.0.1
